+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 v0.4.0
 - Implemented the `exponential::integral` special function
+- Implemented the `Cauchy` (otherwise known as the `Lorenz`) distribution
 
 v0.3.2
 - Implemented the `FisherSnedecor` (F) distribution